Move state related structs in the client from block to state
