+++ crates/starknet_client/src/objects/mod.rs
@@ -2,6 +2,7 @@ pub mod block;
 #[cfg(test)]
 mod block_test;
 pub mod deprecated_contract_class;
+pub mod state;
 pub mod transaction;
 #[cfg(test)]
 mod transaction_test;
